fix(bindings): port to nanobind v3
- Drop the deprecated size argument from NB_TRAMPOLINE in the
  MeshReader/XMLVTK trampolines (nanobind 3 moved the trampoline
  cache to the type object).
- Widen the flags parameter of the custom type casters from
  std::uint8_t to std::uint32_t to match the new from_python
  signature in nanobind 3.

+++ include/pytnl/iostream_caster.h
@@ -42,7 +42,7 @@ public:
    }
 
    bool
-   from_python( handle src, std::uint8_t flags, cleanup_list* cleanup )
+   from_python( handle src, std::uint32_t flags, cleanup_list* cleanup )
    {
       if( getattr( src, "read", none() ).is_none() ) {
          return false;
@@ -109,7 +109,7 @@ public:
    }
 
    bool
-   from_python( handle src, std::uint8_t flags, cleanup_list* cleanup )
+   from_python( handle src, std::uint32_t flags, cleanup_list* cleanup )
    {
       if( getattr( src, "write", none() ).is_none() ) {
          return false;
